+++ frontend/parameters.config.js
@@ -40,21 +40,7 @@ function(toast, parametersService, cleepService, $timeout) {
         self.$onInit = function() {
             cleepService.getModuleConfig('parameters')
                 .then(function(config) {
-                    self.updateConfig(config);
-
-                    // init leaflet
-                    angular.extend($scope, {
-                        cleepposition: {
-                            lat: self.position.latitude,
-                            lng: self.position.longitude,
-                            zoom: 8
-                        },
-                        cleepdefaults: {
-                            scrollWheelZoom: false,
-                            minZoom: 5,
-                            maxZoom: 12
-                        }
-                    });
+                    self.updateConfig(config, true);
                 });
         };
 
@@ -75,19 +61,19 @@ function(toast, parametersService, cleepService, $timeout) {
         /**
          * Position
          */
-        self.setPosition = function() {
+        self.setPosition = function(latitude, longitude) {
             // check values
-            if( !$scope.cleepposition || !$scope.cleepposition.lat || !$scope.cleepposition.lng ) {
+            if (!latitude || !longitude) {
                 toast.info('Please select position before setting position');
                 return;
             }
-            if( $scope.cleepposition.lat==self.position.latitude && $scope.cleepposition.lng==self.position.longitude ) {
+            if (latitude === self.position.latitude && longitude === self.position.longitude) {
                 toast.info('Position not changed');
                 return;
             }
 
             toast.loading('Setting localisation...');
-            parametersService.setPosition($scope.cleepposition.lat, $scope.cleepposition.lng)
+            parametersService.setPosition(latitude, longitude)
                 .then(() => cleepService.reloadModuleConfig('parameters'))
                 .then(function(config) {
                     self.updateConfig(config);
@@ -98,8 +84,8 @@ function(toast, parametersService, cleepService, $timeout) {
         /**
          * Auth
          */
-        self.toggleAuth = function() {
-            const call = !self.auth.enabled ? parametersService.disableAuth : parametersService.enableAuth;
+        self.setAuth = function (enabled) {
+            const call = enabled ? parametersService.enableAuth : parametersService.disableAuth;
             call().then(() => cleepService.reloadModuleConfig('parameters'))
                 .then((config) => {
                     self.updateConfig(config);
@@ -111,7 +97,7 @@ function(toast, parametersService, cleepService, $timeout) {
             parametersService.deleteAuthAccount(account)
                 .then(() => cleepService.reloadModuleConfig('parameters'))
                 .then((config) => {
-                    self.updateConfig(config);
+                    self.updateConfig(config, true);
                     toast.success('Account deleted');
                 });
         };
@@ -120,7 +106,7 @@ function(toast, parametersService, cleepService, $timeout) {
             parametersService.addAuthAccount(self.newAccount, self.newPassword)
                 .then(() => cleepService.reloadModuleConfig('parameters'))
                 .then((config) => {
-                    self.updateConfig(config);
+                    self.updateConfig(config, true);
                     self.newAccount = '';
                     self.newPassword = '';
                     toast.success('Account created');
@@ -130,14 +116,26 @@ function(toast, parametersService, cleepService, $timeout) {
         /**
          * Update controller config
          */
-        self.updateConfig = function(config) {
+        self.updateConfig = function(config, reloadAccounts=false) {
             self.position = config.position;
             self.hostname = config.hostname;
             self.sun = config.sun;
             self.country = config.country;
             self.timezone = config.timezone;
             self.auth.enabled = config.authenabled;
-            cleepService.syncVar(self.auth.accounts, config.authaccounts);
+
+            if (reloadAccounts) {
+                self.auth.accounts.splice(0, self.auth.accounts.length);
+                for (const account of config.authaccounts) {
+                    self.auth.accounts.push({
+                        icon: 'account',
+                        title: account,
+                        clicks: [
+                            { icon: 'delete', tooltip: 'Delete', click: self.deleteAuthAccount, meta: { account } },
+                        ],
+                    });
+                }
+            }
         };
 
     }];
@@ -147,7 +145,7 @@ function(toast, parametersService, cleepService, $timeout) {
         replace: true,
         scope: true,
         controller: parametersController,
-        controllerAs: 'parametersCtl',
+        controllerAs: '$ctrl',
     };
 }]);